Version 4 main features added * all population data ends with _pop[1-9]* This was needed to facilitate processing with get_indicator2data().
* Question to get taxonID for GBIF, NCBI and local databases were added

* Numeric questions on the number of populations, Ne, Nc, lower and upper bounds were added validation constraints so that they have to be >0 or -999, and that the lower and upper bounds can’t be larger and lower than Nc or Ne, respectively. This will prevent adding the lower bound in the upper question, for instance.

* Population 2-100 data autofills with the answers of Population 1, in these questions:

What is the origin of population 2?
In what year was the data used to estimate Ne collected for population 2?
What genetic markers were used to estimate Ne for THIS population?
If 'Other', please explain.
What method was used in calculating Ne for THIS population?
Please list all relevant sources reporting on effective population size for THIS population
If there is a census population size (Nc) available, is this a range or point estimate?
In what year was the Nc estimated for population 2?
How was Nc calculated in THIS population?

Minor:

* Population questions on Ne, Nc and their upper/lower bounds are not mandatory now.