ask for pops only if template is not selected
After the previous commit changes there appeared in which population data was being asked in the kobo form even if the template was selected. This was correcting by replacing:

```
(${kobo_tabular} = 'kobo' or ${kobo_tabular} = '' or ${pop_tabular_file} != '') and ${n_extant_populations} >=
```

for:

```
(${kobo_tabular} = 'kobo' or ${pop_tabular_file} != '') and ${n_extant_populations} >=
```

in the population questions.